% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Apply and document inclusion/exclusion criteria}
\usage{
filter_data(data, ...)
}
\arguments{
\item{data}{a \code{data.frame} containing a dataset to be filtered}

\item{...}{functions or lists of functions to be applied to \code{data}. Each
function must be able to take \code{data} as an argument and remove
observations (rows) failing a user-defined exclusion criteria}
}
\value{
A \code{list} containing the filtered dataset, a criteria-specific
report, and a phase-specific report
}
\description{
This function is designed to assist in the application and documentation
of inclusion and exclusion criteria to clinical or epidemiologic datasets.
It allows the analyst to define a series of functions, each corresponding
to one criteria, and then apply them to the dataset.
It is expected that in many analyses certain inclusion/exclusion criteria
or groups of criteria may be applied sequentially. This function
allows analysts to specify filtering functions in \code{...} in the order
that they should be applied. The subsequent output of \code{filter_data}
is a \code{list} containing the newly filtered
dataset and a report of total observations removed for each criteria.
}
\details{
The input of this function is an unfiltered dataset (\code{data}) and
a series of functions or list(s) of functions accepted by \code{...}. Each
function specified in \code{...} should apply one inclusion/exclusion
criteria to \code{data}. Functions specified in
\code{...} should take \code{data} as the input, remove observations (rows)
that failed to meet a certain inclusion criteria (or which meet a certain
exclusion criteria), and return a \code{data.frame} of the same or
fewer rows as \code{data}.

Each functions may be supplied to \code{...} as either 1) a single
function; or 2) as a list of functions. Each argument captured by
\code{...} will be treated as a separte phase of the data-filtering process,
with each phase being applied sequentially to the dataset. This may affect
the number of observations listed as failing individual inclusion/exclusion
criteria within reports contained in \code{filter_data}'s output. It will
not affect the the filtered dataset generated in \code{filter_data}'s output.

Several criteria may be applied simultaneously as part of a single filtering
"phase". To do this, multiple functions may be specified within a \code{list}
captured by \code{...}. Functions supplied directly to \code{...}
will be treated as their own individual phase in the data-filtering process.

Descriptions of inclusion/exclusion criteria may be specified by supplying
functions to \code{...} as named arguments or named lists.
}
\examples{
\dontrun{
data <- data.frame(A = 1:10, B = LETTERS[1:10],
                   stringsAsFactors = FALSE)
filter_data(data,
            list(remove_A_equals_2 = function(x) x[x[, 1] != 2, ],
                 remove_A_equals_8 = function(x) x[x[, 1] != 8, ]),
            list(remove_B_equals_E = function(x) x[x[, 2] != "E", ]))
}
}
